Move pipes and services into folders and Refactor imports
* Fixes #8

* Fixes #13

* Change schema-enforcer.service name to record-fixer
