+++ src/app/editor/abstract-tracker/abstract-tracker.component.ts
@@ -1,6 +1,5 @@
 export abstract class AbstractTrackerComponent {
   
-  
   /**
    * Used for track changes in model
    * that is used to generate elements via *ngFor
