+++ src/app/editor/add-field-dropdown/add-field-dropdown.component.ts
@@ -2,19 +2,19 @@ import { Component, Input, Output, EventEmitter } from '@angular/core';
 
 import { DROPDOWN_DIRECTIVES } from 'ng2-bootstrap/ng2-bootstrap';
 
-import { DifferentKeysPipe } from '../different-keys.pipe';
+import { DifferentKeysPipe } from '../shared/pipes';
 
-import { EmptyValueService } from '../empty-value.service'
+import { EmptyValueService } from '../shared/services';
 
 @Component({
   selector: 'add-field-dropdown',
   directives: [DROPDOWN_DIRECTIVES],
   pipes: [DifferentKeysPipe],
   providers: [EmptyValueService],
-  template: require('./add-field-dropdown.component.html'),
   styles: [
     require('./add-field-dropdown.component.scss')
-  ]
+  ],
+  template: require('./add-field-dropdown.component.html')
 })
 export class AddFieldDropdownComponent  {
