--- src/app/editor/pair.ts
@@ -1,5 +1,0 @@
-export class Pair<T> {
-  constructor(public key: string, public value: T) {
-    
-  }
-}
