+++ src/app/editor/shared/pipes/index.ts
@@ -1,0 +1,3 @@
+export { DifferentKeysPipe } from './different-keys.pipe';
+export { MapToIterablePipe } from './map-to-iterable.pipe';
+export { UnderscoreToSpacePipe } from './underscore-to-space.pipe';
