+++ src/app/editor/shared/services/index.ts
@@ -1,0 +1,5 @@
+export { EmptyValueService } from './empty-value.service';
+export { JsonUtilService } from './json-util.service';
+export { RecordService } from './record.service';
+export { RecordFixerService } from './record-fixer.service';
+export { SchemaRefResolverService } from './schema-ref-resolver.service';
