+++ src/app/editor/shared/services/record-fixer.service.ts
@@ -1,22 +1,23 @@
 import { Injectable } from '@angular/core';
 
 @Injectable()
-export class SchemaEnforcerService {
+export class RecordFixerService {
 
   /**
-   * Enforces schema rules to the record
+   * Fixes the record for the better UI appearance.
+   * Makes use of schema to get types
    */
-  enforceSchema(record: {}, schema: {}) {
+  fixRecord(record: {}, schema: {}) {
     Object.keys(schema).filter(field => record[field])
       .forEach(field => {
+        // TODO: use getType
         if (schema[field].type === 'array' && schema[field].items.properties) {
           // Check if the array field has any property that is also an array. (Arrayception)
           let subSchema = schema[field].items.properties;
           let hasAnyArrayFields = Object.keys(subSchema)
             .some(subField => subSchema[subField].type === 'array');
           if (!hasAnyArrayFields) {
-            // This field will be represented as a table.
-            // So insert empty values for missing subfields to fix table.
+            // This field will be represented as a table.3
             this.insertEmptyForMissingFields(record[field], subSchema);
           }
         }
